pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
authors = [{name = "Repronim developers", email = "info@repronim.org"}]
dependencies = [
    "click",
    "etelemetry",
    "pyshacl",
    "PyLD",
    "requests",
    "requests_cache",
    "pyyaml",
    "beautifulsoup4",
    "lxml",
    "pydantic >= 2.0",
    "pandas",
    "fhir.resources>=v8.0.0"
]
description = "Reproschema Python library"
# Version from setuptools_scm
dynamic = ["version"]
license = {text = "Apache License, 2.0"}
maintainers = [{name = "Repronim developers", email = "info@repronim.org"}]
name = "reproschema"
readme = "README.md"
requires-python = ">=3.10"

[project.optional-dependencies]
all = ["reproschema[doc]", "reproschema[test]"]
dev = ["reproschema[doc]", "reproschema[test]", "black", "pre-commit"]
doc = [
    "packaging",
    "sphinx >= 2.1.2",
    "sphinx_rtd_theme",
    "sphinxcontrib-apidoc ~= 0.3.0",
    "sphinxcontrib-napoleon",
    "sphinxcontrib-versioning"
]
docs = ["reproschema[doc]"]
# For running unit and docstring tests
test = [
    "pytest >= 4.4.0",
    "pytest-cov",
    "pytest-env",
    "pytest-xdist",
    "pytest-rerunfailures",
    "codecov"
]
tests = ["reproschema[test]"]

[project.scripts]
reproschema = "reproschema.cli:main"

[project.urls]
Homepage = "https://github.com/ReproNim/reproschema-py"

[tool.black]
line-length = 79

[tool.codespell]
exclude-file = "CHANGELOG.md"
ignore-words = "codespell_ignore_words.txt"
skip = "./.git"

[tool.hatch.build.hooks.vcs]
version-file = "reproschema/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["reproschema"]

[tool.hatch.version]
source = "vcs"

[tool.isort]
combine_as_imports = true
line_length = 79
profile = "black"
skip_gitignore = true

[tool.pytest.ini_options]
addopts = "-ra --strict-config --strict-markers --doctest-modules --showlocals -v"
doctest_optionflags = "NORMALIZE_WHITESPACE ELLIPSIS"
junit_family = "xunit2"
minversion = "6.0"
xfail_strict = true
